Update Remediation.ps1
Updated to use the same "LanguageResources" registry branch for Office detection, as per the discovery script.
Also added an additional line (240) to rebuild the LocalRegistryProviders array since changes could have been made in the Providers foreach loop since the last time this variable was created. Without this change, the spotlight\providers key will get deleted again on a default installation where this key does not exist and the LocalRegistryProviders is empty.